PushbackTokenizer: Range is in long, so use Long.MAX_VALUE instead of Integer.MAX_VALUE
git-svn-id: http://josm.openstreetmap.de/svn/trunk@12363 0c6e7542-c601-0410-84e7-c038aed88b3b

+++ src/org/openstreetmap/josm/actions/search/PushbackTokenizer.java
@@ -332,7 +332,7 @@ public class PushbackTokenizer {
                 return new Range(0, Math.abs(currentNumber));
             }
         } else if (isRange && currentRange == null) {
-            return new Range(currentNumber, Integer.MAX_VALUE);
+            return new Range(currentNumber, Long.MAX_VALUE);
         } else if (currentNumber != null && currentRange != null) {
             return new Range(currentNumber, currentRange);
         } else {
